Bug fixes and cleanup 1
